use CreateCosignerShares for sharding
